+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [1.24.3] - 2026-05-20
 ### Added
-- PCI Lookup API for unidentified GPUs, so 'Choose GPU' dialog for multi-GPU setups don't show "Generic Driver" texts
+- Re-implemented PCI Lookup API for unidentified GPUs, so 'Choose GPU' dialog for multi-GPU setups don't show "Generic Driver" texts
 
 ### Fixed
 - Crash if GPU has a letter in its device ID (such as RTX 2070 SUPER)